requirements.txt
Flask==3.0.1
Flask-MySQL==1.5.2
Flask-MySQLdb==2.0.0
APScheduler==3.10.4
